organize controllers and delete obsolete functions
