+++ controllers/auth/auth.js
@@ -1,4 +1,4 @@
-import User from '../models/user.js';
+import User from '../../models/user.js';
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 import { newToken, emailTokenVerification, unverifiedEmailUser } from './emailVerification.js';
