+++ controllers/auth/emailVerification.js
@@ -1,5 +1,5 @@
 'use strict';
-import User from '../models/user.js';
+import User from '../../models/user.js';
 import jwt from 'jsonwebtoken';
 import sgMail from '@sendgrid/mail';
