+++ controllers/calendar/availability.js
@@ -1,14 +1,11 @@
-import User from '../models/user.js'
+import User from '../../models/user.js'
 // import methods from availability util as needed
-
-
 // Potential Availability Controllers
-//
-// Get Single User single day availability
-// Get Single User full week availability
-// Get Full Team availability for single day
-// Get Full Team Availability for a full week
-// Get Common Availability for a team
+  // Get Single User single day availability
+  // Get Single User full week availability
+  // Get Full Team availability for single day
+  // Get Full Team Availability for a full week
+  // Get Common Availability for a team
 
 export const getUserDayAvailability = async (req, res) => {
     try {
