+++ controllers/chats.js
@@ -1,5 +1,3 @@
 // Potential Chat Controllers
-//
-// Create New Chat
-// Get All Chats for a Project
-// 
+  // Create New Chat
+  // Get All Chats for a Project
